+++ lib/features/trips/trip_list/presentation/bloc/trip_list_bloc.dart
@@ -5,6 +5,7 @@ import 'package:delivery_app/core/error/failures.dart';
 import 'package:delivery_app/core/network/network_status.dart';
 import 'package:delivery_app/core/usecase/usecase.dart';
 import 'package:delivery_app/features/trips/shared/domain/entities/trip_entity.dart';
+import 'package:delivery_app/features/trips/shared/domain/entities/trip_extensions.dart';
 import 'package:delivery_app/features/trips/shared/domain/usecases/trip_usecases.dart';
 
 part 'trip_list_event.dart';